Refactor database handling
